trainer/requirements.txt
django
django-ninja>=0.22.0
django-ratelimit
python-dotenv
channels
channels_redis
daphne
watchfiles
psutil

transformers
numpy
trl>=0.9.4,<1.0.0
peft
accelerate>=0.32.0
datasets
bitsandbytes
scipy
sentencepiece
tensorboard